build: create token build styles
