+++ packages/library/scripts/build-styles.mjs
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { createTokens } from './style-dictionary-create.mjs';
+
+createTokens();
